ai/requirements.txt
torch==2.2.2
torchvision==0.17.2
mediapipe==0.10.14
opencv-python==4.9.0.80
numpy<2
matplotlib==3.8.3
Pillow==10.2.0
scikit-learn==1.4.1
kagglehub==0.2.5